Noved TelegramDCStatus to types folder
